requirements.txt
Django==2.1.5
djangorestframework==3.9.0
psycopg2-binary==2.7.6.1
factory-boy==2.11.1
ipdb==0.11
django-extensions==2.1.4
